Add an IndependentController
